refactor(request): extract pull request branch update call
